requirements.txt
aiogram==3.12.0
asyncio==3.4.3
asyncpg==0.29.0
python-dotenv==1.0.1
requests==2.32.3